requirements.txt
Django==4.0.3
asgiref==3.5.0
boto3==1.24.67
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
firebase-admin==5.2.0
Pillow==9.2.0
PyJWT==2.4.0
mysqlclient==2.1.1
requests==2.28.1
six==1.16.0
django-filter==22.1
redis==4.0.2
channels==3.0.5
channels-redis==3.4.0
redis-server==6.0.9
xlrd==1.2.0